keyinput: Merge keyparser into modeparsers
