+++ qutebrowser/keyinput/modeman.py
@@ -25,7 +25,7 @@ import attr
 from PyQt5.QtCore import pyqtSlot, pyqtSignal, Qt, QObject, QEvent
 from PyQt5.QtWidgets import QApplication
 
-from qutebrowser.keyinput import modeparsers, keyparser
+from qutebrowser.keyinput import modeparsers
 from qutebrowser.config import config
 from qutebrowser.commands import cmdexc, cmdutils
 from qutebrowser.utils import usertypes, log, objreg, utils
@@ -67,11 +67,11 @@ def init(win_id, parent):
     keyparsers = {
         KM.normal: modeparsers.NormalKeyParser(win_id, modeman),
         KM.hint: modeparsers.HintKeyParser(win_id, modeman),
-        KM.insert: keyparser.PassthroughKeyParser(win_id, 'insert', modeman),
-        KM.passthrough: keyparser.PassthroughKeyParser(win_id, 'passthrough',
+        KM.insert: modeparsers.PassthroughKeyParser(win_id, 'insert', modeman),
+        KM.passthrough: modeparsers.PassthroughKeyParser(win_id, 'passthrough',
                                                        modeman),
-        KM.command: keyparser.PassthroughKeyParser(win_id, 'command', modeman),
-        KM.prompt: keyparser.PassthroughKeyParser(win_id, 'prompt', modeman),
+        KM.command: modeparsers.PassthroughKeyParser(win_id, 'command', modeman),
+        KM.prompt: modeparsers.PassthroughKeyParser(win_id, 'prompt', modeman),
         KM.yesno: modeparsers.PromptKeyParser(win_id, modeman),
         KM.caret: modeparsers.CaretKeyParser(win_id, modeman),
         KM.set_mark: modeparsers.RegisterKeyParser(win_id, KM.set_mark,
